docs(guide): Edit setup instructions to be as minimal as possible
Removes the setup with `bun create cloudflare` and shows you how to set up a `wrangler.jsonc` with the minimum required keys. This might be useful for people who have never deployed to cloudflare. By the way for me deploying is always the **first** step in creating my discord bot because I'm not going to spend a lot of time making something that works locally, but not being sure if it actually works when deployed. That's why I think it's good to assume that people might not have a project yet at this stage.